Adding more tests and splitting module
